+++ src/test/kotlin/tornadofx/testapps/AccelTest.kt
@@ -1,0 +1,25 @@
+package tornadofx.testapps
+
+import javafx.scene.control.Alert
+import javafx.scene.input.KeyCode
+import javafx.scene.input.KeyCodeCombination
+import javafx.scene.input.KeyCombination
+import tornadofx.*
+
+class AccelTestApp : App(AccelTest::class)
+
+class AccelTest : View() {
+    override val root = stackpane {
+        add(AccelView::class)
+    }
+}
+
+class AccelView : View() {
+    override val root = button("Click me") {
+        accelerator(KeyCodeCombination(KeyCode.A, KeyCombination.SHORTCUT_ANY, KeyCombination.SHIFT_ANY))
+        action {
+            alert(Alert.AlertType.INFORMATION, "Fire!", "You clicked.")
+        }
+    }
+
+}