+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.7.1-SNAPSHOT]
 
+- Button.accelerator also works when button is embedded in sub view (https://github.com/edvin/tornadofx/issues/253)
 - DataGrid correctly calculates horizontal scrollbar
 - DataGrid.maxRows will constrain the max number of rows and override maxCellsInRow if needed (https://github.com/edvin/tornadofx/issues/287)
 - DataGrid properties are now StylableObjectProperties to make them bindable